Trait cmpmore::CmpMore

pub trait CmpMore: PartialOrd + Sized {
    // Provided methods
    fn change_min(&mut self, rhs: Self) { ... }
    fn change_max(&mut self, rhs: Self) { ... }
}

Provide method versions of change_min(), change_max()

Provided Methods

fn change_min(&mut self, rhs: Self)

If self is larger than rhs, override self by rhs.

Examples

let mut x: i32 = 3;
x.change_min(2);
assert_eq!(x, 2);

fn change_max(&mut self, rhs: Self)

If self is smaller than rhs, override self by rhs.

Examples

let mut x: i32 = 3;
x.change_max(4);
assert_eq!(x, 4);

Object Safety

This trait is not object safe.